HSH-131 Added Doctors validation

+++ src/app/component/visit-summary/additional-comment/additional-comment.component.ts
@@ -10,10 +10,7 @@ import {
   animate,
   keyframes,
 } from "@angular/animations";
-import { MatSnackBar } from "@angular/material/snack-bar";
-declare var getEncounterProviderUUID: any,
-  getFromStorage: any,
-  getEncounterUUID: any;
+declare var getEncounterUUID: any;
 
 @Component({
   selector: "app-additional-comment",
@@ -58,9 +55,8 @@ export class AdditionalCommentComponent implements OnInit {
   constructor(
     private service: EncounterService,
     private diagnosisService: DiagnosisService,
-    private snackbar: MatSnackBar,
     private route: ActivatedRoute
-  ) {}
+  ) { }
 
   ngOnInit() {
     this.visitUuid = this.route.snapshot.paramMap.get("visit_id");
@@ -80,9 +76,7 @@ export class AdditionalCommentComponent implements OnInit {
     const date = new Date();
     const form = this.commentForm.value;
     const value = form.comment;
-    const providerDetails = getFromStorage("provider");
-    const providerUuid = providerDetails.uuid;
-    if (providerDetails && providerUuid === getEncounterProviderUUID()) {
+    if (this.diagnosisService.isSameDoctor()) {
       this.encounterUuid = getEncounterUUID();
       const json = {
         concept: this.conceptComment,
@@ -95,17 +89,15 @@ export class AdditionalCommentComponent implements OnInit {
         this.diagnosisService.isVisitSummaryChanged = true;
         this.comment.push({ uuid: resp.uuid, value: value });
       });
-    } else {
-      this.snackbar.open("Another doctor is viewing this case", null, {
-        duration: 4000,
-      });
     }
   }
 
   delete(i) {
-    const uuid = this.comment[i].uuid;
-    this.diagnosisService.deleteObs(uuid).subscribe((res) => {
-      this.comment.splice(i, 1);
-    });
+    if (this.diagnosisService.isSameDoctor()) {
+      const uuid = this.comment[i].uuid;
+      this.diagnosisService.deleteObs(uuid).subscribe((res) => {
+        this.comment.splice(i, 1);
+      });
+    }
   }
 }
