+++ src/app/component/visit-summary/diagnosis/diagnosis.component.ts
@@ -10,10 +10,7 @@ import {
   animate,
   keyframes,
 } from "@angular/animations";
-import { MatSnackBar } from "@angular/material/snack-bar";
-declare var getEncounterProviderUUID: any,
-  getFromStorage: any,
-  getEncounterUUID: any;
+declare var  getEncounterUUID: any;
 
 @Component({
   selector: "app-diagnosis",
@@ -61,9 +58,8 @@ export class DiagnosisComponent implements OnInit {
   constructor(
     private service: EncounterService,
     private diagnosisService: DiagnosisService,
-    private snackbar: MatSnackBar,
     private route: ActivatedRoute
-  ) {}
+  ) { }
 
   ngOnInit() {
     this.visitUuid = this.route.snapshot.paramMap.get("visit_id");
@@ -90,9 +86,7 @@ export class DiagnosisComponent implements OnInit {
   onSubmit() {
     const date = new Date();
     const value = this.diagnosisForm.value;
-    const providerDetails = getFromStorage("provider");
-    const providerUuid = providerDetails.uuid;
-    if (providerDetails && providerUuid === getEncounterProviderUUID()) {
+    if (this.diagnosisService.isSameDoctor()) {
       this.encounterUuid = getEncounterUUID();
       const json = {
         concept: this.conceptDiagnosis,
@@ -106,17 +100,15 @@ export class DiagnosisComponent implements OnInit {
         this.diagnosisList = [];
         this.diagnosis.push({ uuid: resp.uuid, value: json.value });
       });
-    } else {
-      this.snackbar.open("Another doctor is viewing this case", null, {
-        duration: 4000,
-      });
     }
   }
 
   delete(i) {
-    const uuid = this.diagnosis[i].uuid;
-    this.diagnosisService.deleteObs(uuid).subscribe((res) => {
-      this.diagnosis.splice(i, 1);
-    });
+    if (this.diagnosisService.isSameDoctor()) {
+      const uuid = this.diagnosis[i].uuid;
+      this.diagnosisService.deleteObs(uuid).subscribe((res) => {
+        this.diagnosis.splice(i, 1);
+      });
+    }
   }
 }
